CRUD H10

+++ GanaderiaLasDelicias/Controllers/SchedulesController.cs
@@ -1,6 +1,9 @@
-﻿using System.Linq;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using GanaderiaLasDelicias.Models;
 
@@ -18,13 +21,15 @@ namespace GanaderiaLasDelicias.Controllers
         // GET: Schedules
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Schedules.ToListAsync());
+              return _context.Schedules != null ? 
+                          View(await _context.Schedules.ToListAsync()) :
+                          Problem("Entity set 'SGGContext.Schedules'  is null.");
         }
 
         // GET: Schedules/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Schedules == null)
             {
                 return NotFound();
             }
@@ -46,9 +51,11 @@ namespace GanaderiaLasDelicias.Controllers
         }
 
         // POST: Schedules/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ScheduleId,Description,WorkDays,WorkHours,OffDays")] Schedule schedule)
+        public async Task<IActionResult> Create([Bind("ScheduleId,Name,ShiftType,WorkDays,OffDays,WorkHours,IsActive")] Schedule schedule)
         {
             if (ModelState.IsValid)
             {
@@ -62,7 +69,7 @@ namespace GanaderiaLasDelicias.Controllers
         // GET: Schedules/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Schedules == null)
             {
                 return NotFound();
             }
@@ -76,9 +83,11 @@ namespace GanaderiaLasDelicias.Controllers
         }
 
         // POST: Schedules/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ScheduleId,Description,WorkDays,WorkHours,OffDays")] Schedule schedule)
+        public async Task<IActionResult> Edit(int id, [Bind("ScheduleId,Name,ShiftType,WorkDays,OffDays,WorkHours,IsActive")] Schedule schedule)
         {
             if (id != schedule.ScheduleId)
             {
@@ -111,7 +120,7 @@ namespace GanaderiaLasDelicias.Controllers
         // GET: Schedules/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Schedules == null)
             {
                 return NotFound();
             }
@@ -131,15 +140,23 @@ namespace GanaderiaLasDelicias.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Schedules == null)
+            {
+                return Problem("Entity set 'SGGContext.Schedules'  is null.");
+            }
             var schedule = await _context.Schedules.FindAsync(id);
-            _context.Schedules.Remove(schedule);
+            if (schedule != null)
+            {
+                _context.Schedules.Remove(schedule);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool ScheduleExists(int id)
         {
-            return _context.Schedules.Any(e => e.ScheduleId == id);
+          return (_context.Schedules?.Any(e => e.ScheduleId == id)).GetValueOrDefault();
         }
     }
 }
